+++ CHANGELOG.md
@@ -14,6 +14,8 @@
 
 ### Fixed
 
+- `readme-badges-generator` analysis badges
+
 ### Updated
 
 ## [0.1.0-rc.32] - 2021-12-26